stackoverflow/requirements.txt
Flask
flask-restful
pytz
flask-sqlalchemy
flask-migrate
marshmallow-sqlalchemy
alembic
psycopg2-binary
flask-marshmallow
marshmallow
pyjwt
marshmallow-dataclass
passlib
pytest

